Move Bug to annotations package
